+++ src/hotspot/os/posix/os_posix.hpp
@@ -45,7 +45,7 @@
   } while(((int)_result == OS_ERR) && (errno == EINTR))
 
 #define RESTARTABLE_RETURN_INT(_cmd) do { \
-  int _result; \
+  ssize_t _result; \
   RESTARTABLE(_cmd, _result); \
   return _result; \
 } while(false)
